Moved extension host process handling to its own service
