+++ src/main/main.ts
@@ -15,11 +15,11 @@ import dotnetDataProvider from '@main/services/dotnet-data-provider.service';
 import * as NetworkService from '@shared/services/NetworkService';
 import papi from '@shared/services/papi';
 import { CommandHandler } from '@shared/util/PapiUtil';
-import { fork, spawn } from 'child_process';
 import { ProcessType } from '@shared/globalThis';
 import polyfillLocalStorage from '@node/polyfill/LocalStorage';
 import { resolveHtmlPath } from '@node/util/util';
 import MenuBuilder from './menu';
+import extensionHostService from './services/extension-host.service';
 
 // #region Logging setup
 
@@ -148,7 +148,7 @@ const createWindow = async () => {
     return { action: 'deny' };
   });
 
-  // Remove this if your app does not use auto updates
+  // Removed because it was clogging up the logs. Add this back in if our app ends up using auto updates
   // eslint-disable-next-line
   /* new AppUpdater(); */
 };
@@ -159,6 +159,7 @@ app.on('window-all-closed', () => {
   if (process.platform !== 'darwin') {
     // TODO: cleanly stop the provider (close the ws or send command) - IJH 2022-02-23
     dotnetDataProvider.kill();
+    extensionHostService.kill();
     app.quit();
   }
 });
@@ -166,6 +167,7 @@ app.on('window-all-closed', () => {
 app.on('will-quit', () => {
   // TODO: cleanly stop the provider (close the ws or send command) - IJH 2022-02-23
   dotnetDataProvider.kill();
+  extensionHostService.kill();
 });
 
 // #endregion
@@ -255,50 +257,7 @@ const commandHandlers: { [commandName: string]: CommandHandler } = {
 
 // #region Extension Host
 
-const formatExtensionHostLog = (message: string, tag = '') => {
-  const messageNoEndLine = message.trimEnd();
-  const openTag = `{eh${tag ? ' ' : ''}${tag}}`;
-  const closeTag = `{/eh${tag ? ' ' : ''}${tag}}`;
-  if (messageNoEndLine.includes('\n'))
-    // Multi-line
-    return `${openTag}\n${messageNoEndLine}\n${closeTag}`;
-  return `${openTag} ${messageNoEndLine} ${closeTag}`;
-};
-
-// In production, fork a new process for the extension host
-// In development, spawn nodemon to watch the extension-host
-const sharedArgs = ['--resourcesPath', globalThis.resourcesPath];
-const extensionHost = app.isPackaged
-  ? fork(
-      path.join(__dirname, '../extension-host/extension-host.js'),
-      ['--packaged', ...sharedArgs],
-      {
-        stdio: ['ignore', 'pipe', 'pipe', 'ipc'],
-      },
-    )
-  : spawn(
-      process.platform.includes('win') ? 'npm.cmd' : 'npm',
-      ['run', 'start:extension-host', '--', ...sharedArgs],
-      {
-        stdio: ['ignore', 'pipe', 'pipe'],
-      },
-    );
-
-if (!extensionHost.stderr || !extensionHost.stdout)
-  console.error(
-    "Could not connect to extension host's stderr or stdout! You will not see extension host console logs here.",
-  );
-else if (process.env.IN_VSCODE !== 'true') {
-  // When launched from VSCode, don't re-print the console stuff because it somehow shows it already
-  extensionHost.stderr.on('data', (data) =>
-    console.error(formatExtensionHostLog(data.toString(), 'err')),
-  );
-  extensionHost.stdout.on('data', (data) =>
-    console.log(formatExtensionHostLog(data.toString())),
-  );
-}
-
-extensionHost.on('exit', () => console.warn('extensionHost just exited!'));
+extensionHostService.start();
 
 setTimeout(async () => {
   console.log(
